fix: zee workflow arg passing
