+++ packages/ai-agent-sdk/.env.example
@@ -1,2 +1,8 @@
-OPENAI_API_KEY=
-GOLDRUSH_API_KEY=
+# LLM API Keys
+OPENAI_API_KEY
+GEMINI_API_KEY
+DEEPSEEK_API_KEY
+GROK_API_KEY
+
+# Tool API Keys
+GOLDRUSH_API_KEY
