+++ packages/ai-agent-sdk/package.json
@@ -36,6 +36,8 @@
     },
     "scripts": {
         "test": "vitest",
+        "test:llm": "vitest src/core/llm/llm.test.ts",
+        "test:zee": "vitest src/core/zee/zee.test.ts",
         "build": "tsc",
         "clean": "rm -rf dist",
         "prepublishOnly": "cp ../../README.md . && npm run clean && npm run build",
