+++ packages/ai-agent-sdk/src/core/tools/base.ts
@@ -5,13 +5,13 @@ export class Tool {
     private _schema: AnyZodType;
     private _description: string;
 
-    private _execute: (parameters: any) => Promise<string>;
+    private _execute: (parameters: unknown) => Promise<string>;
 
     constructor(
         id: string,
         description: string,
         schema: AnyZodType,
-        execute: (parameters: any) => Promise<string>
+        execute: (parameters: unknown) => Promise<string>
     ) {
         this.id = id;
         this._description = description;
@@ -27,7 +27,7 @@ export class Tool {
         return this._schema;
     }
 
-    execute(parameters: any) {
+    execute(parameters: unknown) {
         return this._execute(parameters);
     }
 }
@@ -36,7 +36,7 @@ interface ToolOptions {
     id: string;
     description: string;
     schema: AnyZodType;
-    execute: (parameters: any) => Promise<string>;
+    execute: (parameters: unknown) => Promise<string>;
 }
 
 export const createTool = (options: ToolOptions) => {
