+++ packages/ai-agent-sdk/src/core/zee/zee.types.ts
@@ -1,0 +1,7 @@
+import type { Agent } from "../agent";
+
+export type ZeeWorkflowOptions = {
+    description: string;
+    output: string;
+    agents: Record<string, Agent>;
+};
